feat: Gomobile SDK implementation stubs
Added stubs for various SDK functionality. This is intended as a simple starting point to allow for testing of the gomobile bindings. Real functionality still needs to be added, and the method signatures and comments/documentation still need to be worked on. Many of them are just placeholders.

Some other notes:
* Added a Crypto interface.
* For the storage implementation stubs, I added very basic storage functionality that we can try using while testing out the gomobile bindings.
* Updated linter script to lint the wallet-sdk-gomobile folder.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>