fix: Circular dependency
The cmd/wallet-sdk-gomobile module was depending on the cmd/utilities module (with a replace statement), and the cmd/utilities module was depending on cmd/wallet-sdk-gomobile module (with a recent commit tag). This causes import issues when updates are made to the interfaces in cmd/wallet-sdk-gomobile.

This commit fixes that by moving the code from cmd/utilities to cmd/wallet-sdk-gomobile and making it a package instead of a module. In order to avoid the exported types and methods from appearing in the bindings (as they are internal details), I modified the build scripts to exclude that package.

This solution allows us to reuse those wrapper objects while preventing a circular dependency and also prevents them from showing up in the bindings.

Signed-off-by: Derek Trider <Derek.Trider@securekey.com>