feat(sdk): Merge higher-level errors into mobile error details field
This replaces the solution introduced in my previous commit where I added a "Full" field to the mobile error to give the caller a way to get the complete error message. This new solution allows the higher-level error messages to not be lost and also not require the caller to examine the parsed error message any differently.

Signed-off-by: Derek Trider <derek.trider@gendigital.com>