docs(sdk): Remove "last updated" lines
The docs have a "last updated" line containing the date and commit hash of their last update. There are a few problems with this:
* It's virtually impossible to specify the hash of the commit you're working on, since the commit hash relies on the content of your commit. Therefore the best you can do is list a previous commit hash. If you wanted to add docs for some new code you added, then you'd have to commit the code first (to get the hash), and then add the docs afterward, which is inconvenient.
* It's easy to forget to update the date and hard to guarantee the date will match when the pull request actually gets merged in.

For these reasons, and since Git can already tell you the last commit that a file was modified in anyway, it makes sense to simply remove these lines.

Signed-off-by: Derek Trider <derek.trider@gendigital.com>