fix(sdk): Update linter
Some new dependencies from aries-framework-go that we will need have some incompatibilities with the older linter version we're using. Updating the linter fixes this. This commit also updates code to comply with new rules in the linter (except for depguard, which has been added to the list of exceptions for now as it requires more consideration).

Signed-off-by: Derek Trider <derek.trider@gendigital.com>